requirements.txt
tk
awscli
pillow
boto3
ebaysdk
openai
opencv-python
